requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
jinja2==3.1.4
python-dotenv==1.0.1
psycopg2-binary==2.9.10
sqlalchemy==2.0.30
itsdangerous==2.2.0   # セッション管理（管理者ログイン）
python-multipart==0.0.9  # フォームデータ受け取り
